[fs] Fix race creating a connection from a non-FIDL-dispatching thread
When the client end of a fuchsia.io channel closes, or if the client
sends a Node.Close message, the server will attempt to teardown the
server fs::Connection instance. When this happens before the connection
is registered with the Vfs, we would hit an assertion in
fbl::DoublyLinkedList when attempting to remove a connection that has
not been added to the list of connections.

The fix is to register the connection before starting FIDL message
dispatch.

Fixed: 45912
Fixed: 48267

Test: Added connection_integration_tests.cc
Test: https://fuchsia-review.googlesource.com/c/fuchsia/+/374513/9 runs
the flaky test 1000 times on infra and passes (see fuchsia-arm64-release).
Without this change, fuchsia-arm64-release would be red when multiplied
1000 times, as witnessed in
https://fuchsia-review.googlesource.com/c/fuchsia/+/374493.

Change-Id: I29bbac221585b54e82b26ba8be06756b583126e3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/374513
Reviewed-by: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Ricardo Vargas <rvargas@google.com>
Testability-Review: Gianfranco Valentino <gevalentino@google.com>
Testability-Review: Ricardo Vargas <rvargas@google.com>
Commit-Queue: Yifei Teng <yifeit@google.com>