[gn][sdk] Add missing file to archive manifest action's outputs.
The archive manifest action in the GN SDK's fuchsia_package() rule
generates not only the archive manifest, but also the build-Ids
file and a file describing the package version and name.

Include the package version/name file in the |outputs| of the
action, so that it is visible to GN. This is required by Chromium
which uses |data_deps| to have |outputs| of the archive manifest
step visible in |runtime_deps|, to allow isolate archives to be
created with the correct contents.

Bug: 49109, chromium:1050703
Change-Id: Ic154b106529b1f46d97d990f478733001d81030d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/375953
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Wez . <wez@google.com>