[appmgr] Policy checker integration tests.
This changeset adds integration testing for appmgr runtime policy
enforcement. This validates that policies that are not part of the
respective allowlists that attempt to utilize the restricted services
and features are denied at launch.

This initial pass adds a proof of concept with three tests:
1. none.cmx - Just verifies you can retrieve a ping from the test
   component with no special sandbox features or services.
2. package_cache_denied.cmx - Adds "fuchsia.pkg.PackageCache" to its
   sandbox service but is not added to the allowlist.
3. package_cache_allowed.cmx - Adds "fuchsia.pkg.PackageCache" to its
   sandbox service but adds it to the _eng allowlist.

Bug: 48899

Change-Id: Ib71f14e5736c2eeb46d177021b9080c1ede4688b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/375834
Reviewed-by: Benjamin Wright <benwright@google.com>
Reviewed-by: Bryan Henry <bryanhenry@google.com>
Commit-Queue: Benjamin Wright <benwright@google.com>
Testability-Review: Bryan Henry <bryanhenry@google.com>