[netstack3] Make IGMP and MLD more type safe
Make IGMP and MLD more type safe by making it impossible to store a
group which is in the Non-Member state. The Non-Member state is used
in the IGMP and MLD RFCs as a sort of sentinal state when a group is
not yet joined. However, actually implementing this state poses two
problems:
- If it's possible to construct a new GMP state without performing
  any transitions - and the only valid state to start in is the
  Non-Member state - then it's possible to store GMP state for a
  group we aren't actually a member of
- If it's possible to transition an existing GMP state to the Non-
  Member state, then the same problem applies

Concretely, since we store multicast groups in a `HashMap`, having
the Non-Member state be representable in memory means that it's
possible (from the type system's perspective) to have an entry in
that map which is in the Non-Member state. This would be a violation
of our invariants - we only store groups we're actually a member of
- but the type system doesn't know that.

We make two concrete changes:
- The only way to construct a GMP state now requires performing the
  "join group" transition at the time of creation
- The only way to perform the "leave group" transition consumes the
  state by value, ensuring that it can no longer be stored

Test: Existing unit tests in the `ip::gmp`, `ip::gmp::igmp`, and
      `ip::gmp::mld` modules cover the behavior changes. Run with:

      fx test netstack3_core_lib_test

Change-Id: Id4bf394dc8b384d579abbe2eb528078742cd3df3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/376881
Commit-Queue: Joshua Liebow-Feeser <joshlf@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Testability-Review: Tamir Duberstein <tamird@google.com>
Testability-Review: Bruno Dal Bo <brunodalbo@google.com>