[mediaplayer] defer stream sink operation
When a client creates an elementary stream source and asks it to
create a stream sink, that stream sink can't do anything until it's
added to the graph, which happens after the source is set on the
player. If the client calls a method on the stream sink before that
point, the player panics.

The client should be able to queue up messages on the stream sink
before SetSource. This CL enables that by having the stream sink
implementation defer binding its FIDL channel until the sink is
ready to handle the messages. The message will wait in the channel
until that point.

TEST: Added a test that issues message to the stream sink before
      SetSource is called.

MULTIPLY: `[
  {
    "name": "mediaplayer_tests",
    "os": "fuchsia",
    "total_runs": 30
  }
]`

Fixed: 48656

Change-Id: Ied0a6eb0632dae816bd88a109b0c6d64aa7859fb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/376773
Commit-Queue: Austin Foxley <afoxley@google.com>
Reviewed-by: Austin Foxley <afoxley@google.com>
Testability-Review: Austin Foxley <afoxley@google.com>