[sl4f][benchmarks] Remove unnecessary return value from convertResults()
convertResults() now raises an exception on failure, so it's not
necessary for it to have a return value that callers check.  The
Catapult file is not useful to callers because they are not expected
to interpret that file.

Change the comment to better reflect what convertResults() does.

Bug: none
Test: fx run-host-tests sl4f_client_tests
Change-Id: I282004969d2935bf2a3e278ec1ba3de3c0c85f8a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/372563
Reviewed-by: Fadi Meawad <fmeawad@google.com>
Testability-Review: Nathan Rogers <nathanrogers@google.com>
Commit-Queue: Mark Seaborn <mseaborn@google.com>