[build][unification] Don't copy from base to variants
This CL prevents targets from being copied from the default toolchain to
variant toolchains in `variant_target`. The template is designed to do
the opposite: copy from variants to base. However, it's possible to have
a binary as a data_dep of a library needed by a variant-selected target
that is NOT selected by the variant; e.g. libfs-management has a
data_dep on blobfs. In this case, the current `variant_target` copies
the default blobfs into a variant build_dir, leading to a failure in
finalize_manifest.py when blobfs's interp doesn't match the variant's
libprefix.

Bug: 49257
Change-Id: Id5c09ee73e824aa7f021b9c9ee9f403ba4893323
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/377440
Reviewed-by: P.Y. Laligand <pylaligand@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Aaron Green <aarongreen@google.com>