[modular][tests] Fix asan use-after-free in agent_runner_unittest
The TestAgent class, defined in the unittest, was serving a PseudoDir in
order to let clients set a Binding which itself enables serving of a
ServiceNamespace instance, thus
taking a dependency on the Binding and ServiceNamespace instances. The initialization
(and destruction order) of TestAgent allowed the Binding and,
importantly, the ServiceNamespace instance to be destroyed prior to the
PseudoDirServer.

This CL ensures the PseudoDirServer is destroyed first.

Fixed: 49304
Tested: fx test -o agent_runner_unittest -- --gtest_repeat=1000
--gtest_break_on_failure=1

Change-Id: Ic7f3526433a1afa2df063204854b26706b3d0135
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/377449
Reviewed-by: Vickie Cheng <vickiecheng@google.com>
Testability-Review: Vickie Cheng <vickiecheng@google.com>
Commit-Queue: Gabe Schine <thatguy@google.com>