[flatland] Implement SetLinkSize() using scale matrices.
The parent-authored TransformHandle that serves as the link
attachment point now has a scale matrix that performs the scaling
from link size to logical size. The scale is the ratio of link size
logical size.

Since logical size is now the denominator of that scale matrix, this
change also adds an additional restriction on the value for logical
size: they must be positive. Considering logical size effectively
represents a clip box, this restriction makes sense.

The initial logical size is also used as the initial content size.
This is primarily convenient because it results in an identity matrix
on the parent-owned link transform, but also ensures the link is
ready to be rendered as soon as it is attached to a Transform.

TEST=fx test fuchsia-pkg://fuchsia.com/scenic_tests#meta/flatland_unittests.cmx

MULTIPLY: `[
  {
    "name": "scenic_tests",
    "os": "fuchsia",
    "total_runs": 40
  }
]`

Change-Id: I4f16a37c1b4e1e889b342b1e279dbbfdfdb43462
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/375114
API-Review: Adam Gousetis <adamgousetis@google.com>
Testability-Review: Craig Stout <cstout@google.com>
Commit-Queue: Evan Birenbaum <birenbaum@google.com>
Reviewed-by: Adam Gousetis <adamgousetis@google.com>