[weave][adaptation] Report config type mismatch
When the requested configuration value does not match the type that is
requested, we previously returned NOT_FOUND, which is not a proper
description of the error. This change introduces platform-specific error
types and reports are more specific type mismatch error.

Bug: 49159
Test: fx test weave_config_manager_unittests
      fx test configuration_manager_unittests
Change-Id: Idb2a0c124857cb3af804519c7db937fc533bda0d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/377094
Reviewed-by: Neethu Joseph <neethujoseph@google.com>
Testability-Review: Ghanan Gowripalan <ghanan@google.com>
Commit-Queue: Prashanth Swaminathan <prashanthsw@google.com>