[SetUI][core] SwitchboardClient Introduction.
The SwitchboardClient abstracts away access to the Switchboard,
which was previously shared with all clients. This sharing led to the
possibility that the Switchboard's mutex could be acquired and not
released. Since the Switchboard internals rely on acquiring the
mutex themselves, this could lead to a deadlock.

Test: A new test has been added that exercises the code pattern
which can lead to a deadlock in the Switchboard.
fx run-test setui_service_tests -- --test
Fixed: b/151728659

Change-Id: I5e7adad4e04b11712ea6bc9fa6fe79a9aa0562bf