[docs] Remove compartments docs (for now)
Compartments are not yet implemented. We'd like for our docs to
represent the existing reality, not promised.
The current plan of record for compartments is materially different than
what's described here.
When compartments are implemented, we should revisit whether the
explanation of the concept belongs in topology.md or merits its own
page.

Change-Id: Ieeb7bb61143a40124ea5bfa21fc14cf254f8d4de
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/377555
Reviewed-by: Kelly-Ann Dolor <kadolor@google.com>
Reviewed-by: Nick Van der Auwermeulen <nickvander@google.com>
Commit-Queue: Shai Barack <shayba@google.com>