[unpack_debug_symbols] explicitly emit a depfile
Today we rely on a package archive is an input to detect when the action
is a dirty target in need of being run again. Since we want to run the
tool on its in a package-free context (e.g., in the case of a unified
//prebuilt/.build-id directory) this change makes it so the action is
dependent on the provided build-id directory itself.

I tested that jiri (actuall CIPD) does indeed update the modtime of the
directory in updating.

Bug: 41478
Change-Id: I815b720cf016b38c3bd5bb4f0c8ed90343277994
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/377556
Reviewed-by: Jake Ehrlich <jakehehrlich@google.com>
Testability-Review: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>