[e2e] Extract time constants
Extracts two constants from existing code:

(1) The "max elapsed since power on"
(2) The estimate of "reboot to timekeeper start"

Both are used to determine if the system clock was reset to zero on a
reboot (as opposed to reset to zero on power-on only).  If the system
clock is not reset to zero we use an estimate that places system reboot
slightly before the startup of a critical system service,
`timekeeper.cmx`.

The constant for (1) was increased to 3 minutes, since we observed
empirically that some startup sequences extend to around 1 minute mark,
causing successive similar measures to use different estimation methods.
3 minutes should be large enough to encompass all such startup sequences.

Bug: 48749
Change-Id: Idf0b2b32d6345d1b81dc8611780b32800df1cc43
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/376994
Commit-Queue: Filip Filmar <fmil@google.com>
Reviewed-by: Isma Cortes <isma@google.com>
Reviewed-by: mesch <mesch@google.com>
Testability-Review: Filip Filmar <fmil@google.com>