[e2e] Add basemgr startup/shutdown to benchmark
Adds the code that optionally starts and shuts down basemgr in
case it is not already running.

This is required on configurations that do not run basemgr by default,
such as arrested boot.  The command is a no-op on configurations that
already has basemgr running.

Also adds a limited-time wait for measurements, to ensure late
checkpoints are also collected.

Bug: 49269
Change-Id: I94c900b2d4410dc03f77331dac36de0aac327bd7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/377394
Commit-Queue: Filip Filmar <fmil@google.com>
Testability-Review: Filip Filmar <fmil@google.com>
Reviewed-by: Isma Cortes <isma@google.com>