[recovery] Add test to verify serial shell is enabled on recovery-eng.
Since there was a regression where serial was disabled on recovery-eng
it makes sense to prevent that ever occuring again by creating a change
that checks that serial shell is enabled.

This test case simply generates a random string and attempts to echo it
in the shell. It then waits for that message to appear in the logs to
verify that the shell is enabled.

Change-Id: I21cf79f35735967fec0b7daf4887ffa5ee036c15
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/377225
Reviewed-by: James Tucker <raggi@google.com>
Testability-Review: James Tucker <raggi@google.com>
Commit-Queue: Benjamin Wright <benwright@google.com>