[storage] Make use of group IDs more robust.
In my attempt to fix Vmoid leaks, I made a mistake with BlockVmoidDetach
by forgetting to set the group ID, which means it would use group ID 0
which has the potential to collide with something else using that group ID.

To address this, I've moved the assignment of group IDs into client.c
which is the right place for this because it is also where the flag
BLOCKIO_GROUP_ITEM is applied to requests. This will prevent any future
mistakes of this kind; callers no longer need to worry about what group
ID is set to.

This change also improves existing behaviour because we can now support
more then 8 threads accessing the block device, although not more than 8
simulataneously.

Lastly, this addresses an issue where we might not process a response to
a request as soon as it arrives. We would never lose responses, but it
was possible that a response would be delivered with a later
response. For example: let's say there are two threads, A & B, and they
issue requests. It's possible that A would have been blocked until B's
response arrives. This fixes that.

Bug: 48968
Change-Id: Ib02307e9a88fdea89dc41408c7961802bc97f1c2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/377855
Commit-Queue: Chris Suter <csuter@google.com>
Reviewed-by: Gianfranco Valentino <gevalentino@google.com>
Testability-Review: Gianfranco Valentino <gevalentino@google.com>