[graphics][compute][hotsort] Fix the "packed" 64-bit integer comparator
GPU devices that do not support directly comparing 64-bit integers
emulate this operation using the GLSL `usubBorrow()` function on a
uvec2 vector type.

The original implementation was incorrect.

This CL provides both branchless and branchy implementations.

Testability:

The `hotsort-vk-bench` tool is used to gauge performance of the
algorithm and compare its output against a CPU algorithm like
`std::sort`.

The HotSort benchmarking tool was updated to optionally reduce the
range of the generated random numbers.

The tool was also updated to sort a repeated short pattern of dwords.
Note that this is enabled at compile time.

The original faulty comparator can be found in the first patchset in
this CL and removed in the following patchset.

The original comparator was verified to be faulty using the pattern
reported in fxr:371874.

Two new comparator variants were implemented and manually tested
against the example patterns and a range of inputs.

Change-Id: I0871343bfe6e412600a5928d985863ddd8faed7f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/372713
Testability-Review: David Reveman <reveman@google.com>
Commit-Queue: Allan MacKinnon <allanmac@google.com>
Reviewed-by: David Turner <digit@google.com>