[fvm][host]: Allow adding empty minfs partition.
Allocate a partition with DataType guid, 1 slice and
the zxcrypt flag, which considers the extra slices.

The partition has the right guid and magic for minfs, but its
not big enough nor the superblock is correct. The purpose is to
trigger auto format of minfs as the block watcher sees the device.

This should prevent data races of components accessing minfs before
hand.

BUG: 40149

TEST=Manual(TBD).

Change-Id: I334042c986c241775abe38eebf192dd1b376fed0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/367873
Reviewed-by: Richard Chen <tianruichen@google.com>
Reviewed-by: Michael Brunson <mbrunson@google.com>
Testability-Review: Gianfranco Valentino <gevalentino@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>