[kernel][mmu] De-templatize the ArchMmu code
In a previous change an allocation routine that both x86 and arm's page
table code was made a template parameter with the idea that for testing
it could use a different allocator without adding a conditional to the
regular routine.

The idea is sound, but the implementation detail of needing to pass
around this template parameter everywhere plus the amount of code
duplication (around 5KB on arm, somewhere around 12KB on x86) is a bit
excessive.

Pass the test routine in as an argument to ::Init() and have a default
just use the regular routine.

Tested: booted on qemu on arm and x86, with and without KVM. Run system
unit tests as well as kernel unittests, which actually uses the
allocator hook.

Change-Id: Iee2236314fe4d7d49e8e18621cac7beb0410dbde
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/375422
Commit-Queue: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Testability-Review: Nick Maniscalco <maniscalco@google.com>