[kernel][arm64] Spectre V1: Harden syscall entry point
The low-level syscall code on ARM64 roughly does:
  arm64_syscall_dispatcher
  ... (x16 == syscall #, user input)
  1: if (x16 >= #ZX_SYS_COUNT)
  2:   jmp .Lunknown_syscall
  3: x12 <= &.Lsyscall_table
  4: x12 <= x12 + x16 << 4
  5: jmp x12

If x16 is >= ZX_SYS_COUNT but a conditional branch predictor mispredicts
the outcome of the branch at (1), the kernel will speculatively index
into the syscall table based on a user-provided index. This could be
leveraged into an information disclosure gadget for data after
.Lsyscall_table in memory.

Add a conditional select + CSDB to convert out-of-range syscall numbers
to 0; this is after the test / branch at (2), so it will never change
the value of the syscall number in program execution, only in
wrong-path speculative execution.

This was done for x86 in commit eedaf885d6b via an LFENCE.

Bug: 12540 Speculative Execution Mitigations.
Bug: 33667 Spectre mitigations?

Change-Id: I775e5fe92761ee4e7de48dbf8a12fa570cf9e635
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/376883
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Kostya Kortchinsky <kostyak@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Testability-Review: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: Venkatesh Srinivas <venkateshs@google.com>