[build] Introduce build_info API module
The introduction of this file is to provide information for downstream
consumers of build artifacts (like ManagedOS and signing server) to
inspect what 'kind' of build they are looking at (usually to verify that
it is indeed a user build that they are looking at). This is meant to
replace the current look-up mechanism of string matching within args.gn
(which is only meant to be internal GN API).

Bug: 45302
Change-Id: I74b616a876de34e39d7bd7b3ad7fd374aad36088
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/375019
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: James Tucker <raggi@google.com>