[net-types] Extract scopes from multicast IPv6 addresses
The list of IPv6 address scopes - `Ipv6Scope` - now encodes all
scopes that apply to both unicast and multicast addresses. This
allows us to make two changes to the impl of
`ScopeableAddress::scope` for `Ipv6Addr`:
- Extract the scope field from multicast addresses
- Treat site-local unicast addresses as having site-local scope

In order to support these changes, we also introduce support for
the site-local unicast subnet.

Finally, we originally structured scoping using the following enum:

    enum Scope<S> {
        Global,
        NonGlobal(S),
    }

This was based on the idea that only non-global scopes could have
associated zone information, and encoding this in an enum made it
easier for us to enforce this property in `AddrAndZone::new`.
However, it turns out that some non-global scopes are also not
allowed to have zones. Thus, this commit also does away with this
enum in favor of a `Scope` trait with a `can_have_zone` method.

Test: Added `test_ipv6_scope`; updated `test_scoped_address`; run
      with:

      fx test net-types-tests

Change-Id: Id366467234573ecdb949c1ee493087b6c3289e27
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/377539
Commit-Queue: Joshua Liebow-Feeser <joshlf@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Testability-Review: Bruno Dal Bo <brunodalbo@google.com>