[build] Change environment() toolchain name construction to use $os more
The -$os suffix is omitted whenever it's "fuchsia" except for
"host", when it's always included for uniformity (when acting as
host, Fuchsia is not special).  It's also omitted when it matches
the base environment name exactly, e.g. for future "efi".

Note this means that environment_redirect() invocations now need
to pass `os` consistent with the environment() definition in cases
that didn't need that before (though always should have had it).

Bug: 49546
Test: no change to build artifacts
Change-Id: Id2d9933d8f47513eaaea12958555570002a59011
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/378269
Reviewed-by: Petr Hosek <phosek@google.com>
Commit-Queue: Roland McGrath <mcgrathr@google.com>