[fidl] Remove unnecessary complexity
Previously we kept a 'need_send' flag to indicate whether adding more data should cause a wakeup or not.
We already have that data though (by virtue of having a waker or not).

We hid the waker inside an AtomicWaker which was protected by a mutex.

This change removes the redundant flag, exposes the waker as an Option<Waker> to be clear about what's going on.
It seems to also fix at least one 100% CPU usage bug.

Change-Id: Ia8160b99a4178cb8c6335145c378fd1af6f52097
Fixed: 49099
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/378427
Commit-Queue: Craig Tiller <ctiller@google.com>
Reviewed-by: Andrew Davies <awdavies@google.com>
Testability-Review: Craig Tiller <ctiller@google.com>