[kernel][vm] Passively classify page age
This implements passive age classification where we move pages to the
front of the page queue whenever they are looked up. We only do this
for user pager backed pages as a performance optimization, since these
are the only pages that would be eviction candidates, and hence where
the age is relevant.

Being only a passive classifier this does not give accurate accessed
age as once mapped into an address space all future accesses will not
trigger a movement in the page queue. Support for this will be
implemented in a future CL.

Perfcompare summary counts:
  980 test cases in total
  980 test cases had no significant difference (no_sig_diff)
  0 test cases got faster
  0 test cases got slower
  0 test cases added
  0 test cases removed

Bug: 34485

Change-Id: I5f13c4d3c2486a2418f23147ab10f1ae9f1165f7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/360168
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>
Testability-Review: Nick Maniscalco <maniscalco@google.com>