[kernel][arm64] Keep the value of PSTATE.PAN unchanged on taking an exception to EL1
Set Privileged Access Never, on taking an exception to EL1.
0b0 PSTATE.PAN is set to 1 on taking an exception to EL1.
0b1 The value of PSTATE.PAN is left unchanged on taking an exception to EL1.
In a system where the PE resets into EL1, this field resets to an
architecturally UNKNOWN value.

If run fuchsia on Qualcomm Snapdragon-710，the bit is not
set for any processor other than processor 0.When processor 1
runs usrboot and trap into the kernel accessing the userspace
address, it gets stuck.If this bit is set for all processors,
then fuchsia works correctly.Run Fuchsia on the development board
vim2. The initial value of this bit is 1, so there is no problem.

Change-Id: Ia204481ce255bbb4146f28885b1b43d5724cf64c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/353435
Commit-Queue: Carlos Pizano <cpu@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Testability-Review: Corey Tabaka <eieio@google.com>