[virtualization][interrupt] Fix interrupt loss bug
In the original code processing flow of 'gich_maybe_interrupt',
if an interrupt comes when its state in the guest is already active,
it will be dropped. This may reaults in the loss of the last interrupt.
So, we need to ensure that when the state of interrupt is already
active in guest, the new interrupt comes, and we can inject an
interrupt with active and pending state.

Change-Id: I43d7141a31f468c7efdd654c4b0b51ea57f07113
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/373242
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Testability-Review: Abdulla Kamar <abdulla@google.com>
Commit-Queue: Abdulla Kamar <abdulla@google.com>