[sshutil] Move sshclient.go into sshutil
In preparation for deleting most of the duplicated code between
sshclient.go and sshutil.go, I'm moving all of sshclient.go into
the sshutil pacakge, which will be the location for the shared code.
This was mostly a simple copy-paste. I'll actually address the
duplication in a series of follow-up changes.

I also added an OWNERS file for the sshutil package.

Bug: 48042
Change-Id: I53fe11aeb36fc7514d75aca78544d259ea196481
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/378373
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>
Testability-Review: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Oliver Newman <olivernewman@google.com>