[drm] Add fuchsia.media.drm data stores
DRM systems generate data on behalf of clients as part of provisioning
and license management. This data is only usable by the DRM system, but
the client is the owner of the data. The client controls the lifetime of
the data and can select which data set is to be used for a
ContentDecryptionModule. This CL adds new protocols for administering
client DataStores and deprecates the Provisioner protocol.

Bug: 13619

Change-Id: If981f501f7ef34cf369904042748824f394e041f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/358613
Commit-Queue: RJ Ascani <rjascani@google.com>
Testability-Review: Austin Foxley <afoxley@google.com>
Reviewed-by: Wez . <wez@google.com>
API-Review: Wez . <wez@google.com>
API-Review: Dale Sather <dalesat@google.com>