[zircon][fidl][zx] Reland: Extract built-in zx library
This moves the magically injected `library zx` out of fidlc, and into
normal fidl files in //zircon/vdso. This allows the syscall definitions
to share the definition with other fidl, and also to have the rest of
the syscall definitions live in library zx (rather than the fake
"library zz" that was internally mapped to zx by kazoo).

Rather than have every fidl library have to specify a public_deps on zx,
the GN templates are modified to automatically include that dependency,
so previous behaviour is maintained.

This is largely a reland of
https://fuchsia-review.googlesource.com/c/fuchsia/+/343788 other than
updates for various churn that has happened since then. The previous CL
was reverted because of the details of how Chromium generated build
rules (see fxbug.dev/44602), but does not appear to be an issue any more
(based on doing a local Chromium build with tools/fuchsia/local-sdk.py
in the Chromium repo.)

Re-reland, with verify_sdk_api.py patch for transition.

Bug: 7802
Bug: 42717
Bug: 44602

This reverts commit d50813e0a3c4058c4303d6f90fa940f504454898.

Change-Id: I2d13c7dfd5bb81be34f3d2f076d577c8cbffa557
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/378008
Reviewed-by: Pascal Perez <pascallouis@google.com>
API-Review: Pascal Perez <pascallouis@google.com>
Testability-Review: Pascal Perez <pascallouis@google.com>
Commit-Queue: Scott Graham <scottmg@google.com>