[kernel][heap] Add skeleton cmpctmalloc host test
This change ensures that the zircon build compiles the cmpctmalloc
library for host and exports it to the fuchsia build. Moreover, it adds
a dummy host test in the fuchsia build to ensure that everything
actually compiles, here and going forward.

Bug: 49123
Change-Id: I1039190101193db5efbaa0787f8ce0d1040d1a43
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/378271
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Testability-Review: Nick Maniscalco <maniscalco@google.com>