[fidlc] Convert lexer, parser fidlc errors to (Reported)Errors
This builds on the infra added in fxr/374574 and converts all the
reported errors in the lexer and parser to template struct Errors.

It also includes a ReportedError type, which allows for instantiating
and passing around an error before reporting it. ReportedError holds a
pointer to a templated Error definition and a std::tuple of params.
This allows for more flexible error handling, and will in future CL(s)
allow for ErrorReporter to collect ReportedErrors rather than strings.
This will, e.g., allow tests to verify that a particular error with
particular params was reported, rather than doing a string compare.

Error types and formatting logic are moved to error_types.h and error
definitions are left in errors.h.

Test: fidlc host tests
Change-Id: Ib75bf90420e3cf7e98b3b71c8196a2497775eba9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/376934
Testability-Review: Benjamin Prosnitz <bprosnitz@google.com>
Commit-Queue: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Benjamin Prosnitz <bprosnitz@google.com>