[media][tests] log possible cause of flaky result
Under the assumption that this test has been intermittently failing
because another instance of audio_core was still running (thus we
fail to configure the input device; without an input device, all
capturers record silence), this CL explicitly checks for this and
fails with more clear error.

The audio_core_hardware_tests suite is only run on hardware platforms
with built-in audio input devices, as it expressly captures audio and
ensures that it is not digital silence. To isolate intermittent
failures, the test now waits for an input device to be enumerated
before executing. If no device is found in 1 minute, the test fails.

Another possible cause of intermittent failure would be if the capture
gain settings were erroneously attenuating the input signal to the
point of being mistaken for "digital silence". To fix this issue,
now device gain, capture usage volume/gain and capturer stream gain
are all set to unity before testing begins.

MULTIPLY: `[
  {
    "name": "audio_core_hardware_tests",
    "total_runs": 100
  }
]`

Test: test code only
Fixed: 49326

Change-Id: I3f48b6d6bc11d239613cb3b910d15f5848d49209
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/378152
Reviewed-by: Tim Detwiler <tjdetwiler@google.com>
Testability-Review: Tim Detwiler <tjdetwiler@google.com>
Commit-Queue: Martin Puryear <mpuryear@google.com>