[flake] Fix flake in ax88179 test
SIGNAL_STATUS is asserted by both the fake function driver, and by
ethernet when it starts up. As the test can't control the order of
delivery of the interrupt for the fake setting of online (which signals)
and the signal at ethernet startup, check for up to two signals and
confirm that the test gets to "online" after one or both.

Fixed: 49373
Change-Id: Ia9233d9e8120a665097428a104f3885b92299d21
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/378759
Commit-Queue: Scott Graham <scottmg@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Testability-Review: Suraj Malhotra <surajmalhotra@google.com>