[storage] Change how we track changes to the superblock.
We now track changes to the superblock using a dirty flag and then write
the superblock at the end of a transaction if it's dirty.

Previously, it was easy for us to forget to update the superblock. This
change makes it harder to forget and may also provide some efficiency
gains (albeit small) because we will now typically only update the
superblock once per transaction.

Bug: 48842
Change-Id: I6ee607d17c80a095e80a971d23661070849f8d8c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/378314
Commit-Queue: Chris Suter <csuter@google.com>
Testability-Review: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Manali Bhutiyani <manalib@google.com>