[storage] After deallocation, keep blocks reserved until transaction commits.
This is a proper fix for the problem outlined in fxb/37958 (for which a
workaround had already been submitted), and a fix for fxb/45818. The
workaround is removed for minfs. A fix is still required for blobfs.

After deallocating blocks, they cannot be used for data writes until the
transaction that frees them has committed i.e. been sent to the lower
layers.

To simplify things, the splitting/assignment of reservations has been
removed since it adds no additional value. If the number of reserved
blocks is insufficient, then we panic anyway.

Added new tests that prove that fxb/37958 and fxb/45818 are fixed.

Bug: 37958
Bug: 45818
Change-Id: I78959f0531f4dd6af7f17fda7fc395b14cc2c097
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/374278
Commit-Queue: Chris Suter <csuter@google.com>
Reviewed-by: Gianfranco Valentino <gevalentino@google.com>
Testability-Review: Gianfranco Valentino <gevalentino@google.com>