[ulib][fs-management]: 'regsiter_fs' option.
Allows optionally registering with /svc/fuchsia.fshost.Registry.
In test environments, when dealing with IsolatedDevMgr in a packaged
environemnt it introduces failure. This feature is not needed for
testing, or it is used in the platform yet.
Until a proper fix for the registry exists, this allows tests disabling
this option. This would also break isolation, since would provide a test
only to the system wide audience of fshost. This will prevent this from
happening.

// Before this the tests on the pending patch would fail when not in
bringup.

TEST=fx test fvm-tests

Change-Id: I95b913c78cf4de355097e5e79b41012441811771
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/378934
Reviewed-by: Stephen Demos <sdemos@google.com>
Testability-Review: Gianfranco Valentino <gevalentino@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>