[codec_impl] Fix QueueInputPacket() to keep discarded packets as free.
Based on this posted CL from sergeyu@google.com:
https://fuchsia-review.googlesource.com/c/fuchsia/+/372853/3

Previously CodecImpl's packet free/used tracking would get out of sync
with reality if QueueInputPacket() immediately returned the packet due
to the stream being obsolete.  That was causing the codec to fail as
soon as the client tried to reuse the packet that was marked used when
it should have been marked free.

This change fixes the free/used tracking to leave the packet marked
free when the input packet is immediately returned to the client.

Bug: 48489
Change-Id: Iad1408bcbe9c43d7c5091c89e75c96012495311c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/379013
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: John Bauman <jbauman@google.com>
Commit-Queue: Dustin Green <dustingreen@google.com>
Testability-Review: John Bauman <jbauman@google.com>