[storage][volume_image]: VolumeImage schema.
Effectively a defined json for passing the mappings and the volume
information. In the future, if we need to, we may make address either
default to identity (1:1 map) or not required.

For now, our images are always in the FVM, and this is for the fvm tool
as well.

schema:

{
  volume: { volume descriptor }
  address: { address_descriptor }
}

TEST=storage-volume-image-tests

Change-Id: Idc4d6b939ce8075a1ef67bd40e1485b889ce5269
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/378669
Testability-Review: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Richard Chen <tianruichen@google.com>
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>