[audio][sherlock] Add assertion to position notifications
A flake in the AudioDriverTest.OutputNoPositionNotifyAfterStop test
hits an assertion on a position too large for the size of the ring
buffer, geting values 3754061824 vs 385024 for:
EXPECT_LT(ring_buffer_position_, ring_buffer_frames_ * frame_size_);

We add an extra assertions in the driver for a position bigger than
the ring buffer size. This can help the debugging of where the
corruption is occuring.

Bug: 43792

Test: fx run-test audio_device_tests on Sherlock
Change-Id: I2b5e176bdc235216b011ca37d3f9fc3a91331219
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/378538
Reviewed-by: Eric Holland <hollande@google.com>
Testability-Review: Braden Kell <bradenkell@google.com>
Commit-Queue: Andres Oportus <andresoportus@google.com>