[release] archives.json & recipes compatibility stack
This combines commits necessary to be compatible with upcoming
changes to recipes, which will begin consuming archives.json.

[build] Create an archives build API module

These this will soon be deprecated (hopefully). Build targets that
depend on everything should be removed (or realized as some sort of
assembly step). In the meantime, these entries make more sense in their
own API module, instead of being shoehorned into images.json.

A follow-up change will remove the archives from images.json.

[build] Remove unused archives

We are in the process of burning down the list of archives produced by
the infrastructure. This change removes non-tgz variations of
build-archive and the symbol archive; the former were never built by
by the infrastructure and the latter is no longer built.

Change-Id: I83cfadc60f6d2954e7d61d721cec54e74c1a8dbb