[build] Remove unused archives
We are in the process of burning down the list of archives produced by
the infrastructure. This change removes non-tgz variations of
build-archive and the symbol archive; the former were never built by
by the infrastructure and the latter is no longer built.

Test: build-archive.tgz builds and has contents as expected.
Bug: 43568
Change-Id: I866f0199d301e1abbbd5b577b62fb8afdc3ce02c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/378219
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Testability-Review: Roland McGrath <mcgrathr@google.com>