[hypervisor] Add back current thread check
This was accidentally removed when the rest of the pinned thread check
was removed during the recent changes to remove the hypervisor's
reliance on affinity.

We rely on the VCPU to be resumed, or its state to be read/written, only
from the thread it was created on.

Change-Id: I2eda756c28bd63a6a467eb7320d7bf3e6c7a64d6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/379213
Commit-Queue: Abdulla Kamar <abdulla@google.com>
Testability-Review: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>
Reviewed-by: David Greenaway <dgreenaway@google.com>